requirements.txt
# Core screen recording dependencies
mss>=9.0.1                    # High-performance screen capture
opencv-python>=4.8.1.78       # Video processing and encoding
numpy>=1.24.3                 # Array operations
Pillow>=10.0.1                # Image processing

# GUI and window management
pygetwindow>=0.0.9            # Window detection and management

# Audio capture (Windows specific)
sounddevice>=0.4.6            # Audio recording interface
pycaw>=20240210               # Windows audio control

# Note: tkinter is included with Python (no installation needed)
